fix(ecs): 修正BatchRebootSeversOption类命名拼写错误
- 新增正确的类名 BatchRebootServersOption
- 将原有的 BatchRebootSeversOption 标记为 Deprecated（废弃），并继承新类以保持向后兼容
- 修正对应模块和导入语句中的类名拼写错误
- 已通过单元测试验证，旧类名依然兼容，新类名符合规范"
