+++ huaweicloud-sdk-ecs/huaweicloudsdkecs/v2/model/__init__.py
@@ -50,6 +50,7 @@ from huaweicloudsdkecs.v2.model.batch_operate_result_response import BatchOperat
 from huaweicloudsdkecs.v2.model.batch_reboot_servers_request import BatchRebootServersRequest
 from huaweicloudsdkecs.v2.model.batch_reboot_servers_request_body import BatchRebootServersRequestBody
 from huaweicloudsdkecs.v2.model.batch_reboot_servers_response import BatchRebootServersResponse
+from huaweicloudsdkecs.v2.model.batch_reboot_servers_option import BatchRebootServersOption
 from huaweicloudsdkecs.v2.model.batch_reboot_severs_option import BatchRebootSeversOption
 from huaweicloudsdkecs.v2.model.batch_reset_servers_password_request import BatchResetServersPasswordRequest
 from huaweicloudsdkecs.v2.model.batch_reset_servers_password_request_body import BatchResetServersPasswordRequestBody
